+++ Simulation/Aivika/IO/Comp.hs
@@ -1,6 +1,4 @@
 
-{-# LANGUAGE FlexibleInstances, UndecidableInstances #-}
-
 -- |
 -- Module     : Simulation.Aivika.IO.Comp
 -- Copyright  : Copyright (c) 2009-2016, David Sorokin <david.sorokin@gmail.com>
@@ -9,7 +7,7 @@
 -- Stability  : experimental
 -- Tested with: GHC 8.0.1
 --
--- It allows making the 'MonadIO'-based monad an instance of type class 'MonadComp'
+-- It makes the 'IO' monad an instance of type class 'MonadComp'
 -- on top of which the simulation monads can be built.
 --
 module Simulation.Aivika.IO.Comp () where
@@ -22,9 +20,9 @@ import Simulation.Aivika.IO.Generator
 
 import Simulation.Aivika.Trans.Comp
 import Simulation.Aivika.Trans.Exception
-import Simulation.Aivika.Trans.Template
 
--- | A template-based instantiation of the 'MonadComp' type class. 
-instance (Functor m, Monad m, MonadIO m, MonadException m, MonadTemplate m) => MonadComp m where
+-- | An instantiation of the 'MonadComp' type class. 
+instance MonadComp IO where
+-- instance (Functor m, Monad m, MonadIO m, MonadException m, MonadTemplate m) => MonadComp m where
 
   {-# SPECIALISE instance MonadComp IO #-}