+++ Simulation/Aivika/IO/DES.hs
@@ -1,6 +1,4 @@
 
-{-# LANGUAGE FlexibleInstances, UndecidableInstances #-}
-
 -- |
 -- Module     : Simulation.Aivika.IO.DES
 -- Copyright  : Copyright (c) 2009-2016, David Sorokin <david.sorokin@gmail.com>
@@ -9,7 +7,7 @@
 -- Stability  : experimental
 -- Tested with: GHC 8.0.1
 --
--- It allows making the 'MonadIO'-based monad an instance of type class 'MonadDES'
+-- It makes the 'IO' monad an instance of type class 'MonadDES'
 -- used for Discrete Event Simulation (DES).
 --
 module Simulation.Aivika.IO.DES () where
@@ -22,11 +20,11 @@ import Simulation.Aivika.IO.Event
 
 import Simulation.Aivika.Trans.Comp
 import Simulation.Aivika.Trans.DES
-import Simulation.Aivika.Trans.Template
 
 import Simulation.Aivika.IO.QueueStrategy
 
--- | A template-based instantiation of the 'MonadDES' type class.
-instance (Monad m, MonadComp m, MonadIO m, MonadTemplate m, MonadEventQueueTemplate m) => MonadDES m where
+-- | An instantiation of the 'MonadDES' type class.
+instance MonadDES IO where
+-- instance (Monad m, MonadComp m, MonadIO m, MonadTemplate m, MonadEventQueueTemplate m) => MonadDES m where
 
   {-# SPECIALISE instance MonadDES IO #-}