+++ Simulation/Aivika/IO/SD.hs
@@ -1,6 +1,4 @@
 
-{-# LANGUAGE FlexibleInstances, UndecidableInstances #-}
-
 -- |
 -- Module     : Simulation.Aivika.IO.SD
 -- Copyright  : Copyright (c) 2009-2016, David Sorokin <david.sorokin@gmail.com>
@@ -9,7 +7,7 @@
 -- Stability  : experimental
 -- Tested with: GHC 8.0.1
 --
--- It allows making the 'MonadIO'-based monad an instance of type class 'MonadSD'
+-- It makes the 'IO' monad an instance of type class 'MonadSD'
 -- used for System Dynamics (SD).
 --
 module Simulation.Aivika.IO.SD () where
@@ -22,9 +20,9 @@ import qualified Simulation.Aivika.IO.Dynamics.Memo.Unboxed as MU
 
 import Simulation.Aivika.Trans.Comp
 import Simulation.Aivika.Trans.SD
-import Simulation.Aivika.Trans.Template
 
--- | A template-based instantiation of the 'MonadSD' type class.
-instance (Monad m, MonadComp m, MonadIO m, MonadTemplate m) => MonadSD m where
+-- | An instantiation of the 'MonadSD' type class.
+instance MonadSD IO where
+-- instance (Monad m, MonadComp m, MonadIO m, MonadTemplate m) => MonadSD m where
   
   {-# SPECIALISE instance MonadSD IO #-}