+++ Simulation/Aivika/IO/Var.hs
@@ -1,5 +1,5 @@
 
-{-# LANGUAGE TypeFamilies, FlexibleInstances, UndecidableInstances #-}
+{-# LANGUAGE TypeFamilies #-}
 
 -- |
 -- Module     : Simulation.Aivika.IO.Var
@@ -9,7 +9,7 @@
 -- Stability  : experimental
 -- Tested with: GHC 8.0.1
 --
--- The 'MonadIO'-based monad can be an instance 'MonadVar'.
+-- The 'IO' monad is an instance of 'MonadVar'.
 --
 module Simulation.Aivika.IO.Var () where
 
@@ -24,7 +24,6 @@ import Simulation.Aivika.Trans.Internal.Dynamics
 import Simulation.Aivika.Trans.Internal.Event
 import Simulation.Aivika.Trans.Ref
 import Simulation.Aivika.Trans.Signal
-import Simulation.Aivika.Trans.Template
 import Simulation.Aivika.Trans.Var
 
 import Simulation.Aivika.IO.DES
@@ -32,17 +31,18 @@ import Simulation.Aivika.IO.DES
 import qualified Simulation.Aivika.Vector as V
 import qualified Simulation.Aivika.Vector.Unboxed as UV
 
--- | The 'MonadIO' based monad is an instance of 'MonadVar'.
-instance (Monad m, MonadDES m, MonadIO m, MonadTemplate m) => MonadVar m where
+-- | The 'IO' monad is an instance of 'MonadVar'.
+instance MonadVar IO where
+-- instance (Monad m, MonadDES m, MonadIO m, MonadTemplate m) => MonadVar m where
 
   {-# SPECIALISE instance MonadVar IO #-}
 
   -- | A template-based implementation of the variable.
-  data Var m a = 
+  data Var IO a = 
     Var { varXS    :: UV.Vector Double,
           varMS    :: V.Vector a,
           varYS    :: V.Vector a,
-          varChangedSource :: SignalSource m a }
+          varChangedSource :: SignalSource IO a }
 
   {-# INLINABLE newVar #-}
   newVar a =