+++ Simulation/Aivika/Trans.hs
@@ -57,7 +57,6 @@ module Simulation.Aivika.Trans
         module Simulation.Aivika.Trans.Stream,
         module Simulation.Aivika.Trans.Stream.Random,
         module Simulation.Aivika.Trans.Task,
-        module Simulation.Aivika.Trans.Template,
         module Simulation.Aivika.Trans.Transform,
         module Simulation.Aivika.Trans.Transform.Extra,
         module Simulation.Aivika.Trans.Transform.Memo.Unboxed,
@@ -108,7 +107,6 @@ import Simulation.Aivika.Trans.Statistics.Accumulator
 import Simulation.Aivika.Trans.Stream
 import Simulation.Aivika.Trans.Stream.Random
 import Simulation.Aivika.Trans.Task
-import Simulation.Aivika.Trans.Template
 import Simulation.Aivika.Trans.Transform
 import Simulation.Aivika.Trans.Transform.Extra
 import Simulation.Aivika.Trans.Transform.Memo.Unboxed